% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcAllometries.R
\name{allometries_description}
\alias{allometries_description}
\title{Info for allometries}
\usage{
allometries_description(object, ..., id = NULL)
}
\arguments{
\item{object}{\code{lfcAllometries} object, as created by \code{\link{allometries}}}

\item{...}{Expresions resolving to a logical value to filter the allometries. Only
evaluated if \code{id} is NULL.}

\item{id}{Character vector with the allometry/ies id}
}
\value{
a list with the selected allometries and their info
}
\description{
\code{allometries_description} is a wrapper for the \code{$description}
  method of \code{lfcAllometries} objects. See also \code{\link{allometries}}.
}
\examples{

# by id
allomdb <- allometries()
foo <- allometries_description(allomdb, id = "GC_2589")
foo$GC_2589$dependent_var
foo$GC_2589$param_a

# filtering
ht_dn_allometries <- allometries_description(allomdb, dependent_var \%in\% c("GC", "Dn"))
ht_dn_allometries$GC_2589$dependent_var
ht_dn_allometries$GC_2589$param_a

}
\seealso{
Other Allometries functions: \code{\link{allometries_get_data}},
  \code{\link{allometries}}
}
\concept{Allometries functions}
