% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcCatDrought.R
\name{catdrought_get_raster}
\alias{catdrought_get_raster}
\title{Access to the rasters in the Catdrought database}
\usage{
catdrought_get_raster(object, date, resolution, spatial = "stars")
}
\arguments{
\item{object}{\code{lfcCatDrought} object, as created by
\code{\link{catdrought}}}

\item{date}{character with the date of the raster to retrieve, i.e "2020-04-25"}

\item{resolution}{character indicating the desired resolution, can be one of
\code{'200m'}, \code{'1km'} or \code{'smoothed'}}

\item{spatial}{character vector of length 1 indicating the type of raster
object to return, "raster" or "stars", the default.}
}
\value{
A raster object: \code{RasterBrick} if spatial is \code{raster},
  \code{stars} if spatial is \code{stars}. See
  https://r-spatial.github.io/stars/index.html for details about stars
  objects and \code{\link[raster]{raster}} for details about raster objects.
}
\description{
\code{catdrought_get_raster} is a wrapper for the
  \code{$get_raster} method of \code{lfcCatDrought} objects.
  See also \code{\link{catdrought}}.
}
\details{
Connection to database can be slow. Rasters retrieved from the db
  are stored in a temporary cache inside the lfcCatDrought object created by
  \code{\link{catdrought}}, making subsequent calls to the same table are
  faster. But, be warned that in-memory rasters can use a lot of memory!
}
\seealso{
Other catdrought functions: 
\code{\link{catdrought_get_current_time_series}()},
\code{\link{catdrought}()}
}
\concept{catdrought functions}
