% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcNFI.R
\docType{data}
\name{lfcNFI}
\alias{lfcNFI}
\title{importFrom R6 R6Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
lfcNFI
}
\description{
importFrom R6 R6Class
}
\keyword{datasets}
