% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcLiDAR.R
\name{lidar_clip_and_stats}
\alias{lidar_clip_and_stats}
\title{Clip and calculate the mean of raster tables}
\usage{
lidar_clip_and_stats(object, sf, polygon_id_variable, table_names)
}
\arguments{
\item{object}{\code{lfcLiDAR} object, as created by \code{\link{lidar}}}

\item{sf}{sf object with the polygon/s to clip}

\item{table_names}{character vector with the names of the tables to access}
}
\value{
This function returns the same sf object provided with new columns with the
  mean of each polygon for each table requested.
}
\description{
\code{lidar_clip_and_mean} is a wrapper for the \code{$clip_and_mean}
  method of \code{lfcLiDAR} objects. See \code{\link{lidar}}.
}
\examples{
if (interactive()) {
lidardb <- lidar()
}

}
\seealso{
Other LiDAR functions: \code{\link{lidar_avail_tables}},
  \code{\link{lidar_describe_var}},
  \code{\link{lidar_get_lowres_raster}},
  \code{\link{lidar}}
}
\concept{LiDAR functions}
