% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfcLiDAR.R
\name{lidar}
\alias{lidar}
\title{lfcLiDAR class}
\usage{
lidar()
}
\value{
An \code{lfcLiDAR} class object (inherits from \code{\link[R6]{R6Class}}),
  with methods to access the data. See Methods section.
}
\description{
\code{lidar()} creates an object to access the LiDAR database.
}
\section{Methods}{

  \code{lfcLiDAR} objects has one public method:
  \itemize{
    \item{\code{$get_data}: Retrieve and collect LiDAR database rasters. See
          \code{\link{lidar_get_data}} for more details}
  }
}

\examples{
lidardb <- lidar()
lidardb
}
\seealso{
Other LiDAR functions: \code{\link{lidar_avail_tables}},
  \code{\link{lidar_describe_var}},
  \code{\link{lidar_get_data}}
}
\concept{LiDAR functions}
